backend/requirements.txt
opencv-python
mediapipe
numpy
google-generativeai
python-dotenv
fastapi
uvicorn[standard]
python-multipart
